// src/types/promo.ts
export type PromoDataType = {
  color: string;
  conditions: string;
  conditions_path: string;
  delivery: string;
  discount: number;
  end_date: string;
  name: string;
  pic_path: string;
  position: string;
  start_date: string;
};

export type PromosDataType = {
  color: string;
  name: string;
  pic_path: string;
  position: string;
  slug: string;
};